Update RCE autosave when editing in html view
Adds an autosave update in the textarea's onchange handler.
Removes the window unload handler, since it was not getting called anyway

closes LS-1074
flag=rce_auto_save

test plan:
  - follow the recreate steps in the ticket, because
     the syllabus edit page is different from other edit pages
  - on re-editing
    > expect not to get prompted for autosaved content
    > expect the RCE text to be everything you entered last time
  - on any other RCE page (edit assignment, announcement, ...
  - enter some text in the rce
  - flip to html mode and enter some more
  - leave w/o saving (refresh or cancel)
  - edit again
  > expect to be prompted with autosaved content that contains
    everything you entered.
  - flip to html mode and enter some more text
  - save
  > expect the saved item to have all the text
  - edit again
  > expect not to be prompted with autosaved content

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/243011
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>

+++ packages/canvas-rce/src/rce/RCEWrapper.js
@@ -741,9 +741,6 @@ class RCEWrapper extends React.Component {
     if (this.storage) {
       editor.on('change', this.doAutoSave)
       editor.on('blur', this.doAutoSave)
-      window.addEventListener('unload', e => {
-        this.doAutoSave(e)
-      })
 
       this.cleanupAutoSave()
 
@@ -772,16 +769,18 @@ class RCEWrapper extends React.Component {
   // remove any autosaved value that's too old
 
   cleanupAutoSave = (deleteAll = false) => {
-    const expiry = deleteAll
-      ? Date.now()
-      : Date.now() - this.props.autosave.rce_auto_save_max_age_ms
-    let i = 0
-    let key
-    while ((key = this.storage.key(i++))) {
-      if (/^rceautosave:/.test(key)) {
-        const autosaved = this.getAutoSaved(key)
-        if (autosaved && autosaved.autosaveTimestamp < expiry) {
-          this.storage.removeItem(key)
+    if (this.storage) {
+      const expiry = deleteAll
+        ? Date.now()
+        : Date.now() - this.props.autosave.rce_auto_save_max_age_ms
+      let i = 0
+      let key
+      while ((key = this.storage.key(i++))) {
+        if (/^rceautosave:/.test(key)) {
+          const autosaved = this.getAutoSaved(key)
+          if (autosaved && autosaved.autosaveTimestamp < expiry) {
+            this.storage.removeItem(key)
+          }
         }
       }
     }
@@ -813,7 +812,7 @@ class RCEWrapper extends React.Component {
   getAutoSaved(key) {
     let autosaved = null
     try {
-      autosaved = JSON.parse(this.storage.getItem(key))
+      autosaved = this.storage && JSON.parse(this.storage.getItem(key))
     } catch (_ex) {
       this.storage.removeItem(this.autoSaveKey)
     }
@@ -836,35 +835,31 @@ class RCEWrapper extends React.Component {
   }
 
   doAutoSave = (e, retry = false) => {
-    const editor = this.mceInstance()
-    // if the editor is empty don't save
-    if (editor.dom.isEmpty(editor.getBody())) {
-      return
-    }
-    // if no changes have been made,
-    // delete and don't save
-    if (!editor.isDirty()) {
-      this.storage.removeItem(this.autoSaveKey)
-      return
-    }
+    if (this.storage) {
+      const editor = this.mceInstance()
+      // if the editor is empty don't save
+      if (editor.dom.isEmpty(editor.getBody())) {
+        return
+      }
 
-    const content = editor.getContent({no_events: true})
-    try {
-      this.storage.setItem(
-        this.autoSaveKey,
-        JSON.stringify({
-          autosaveTimestamp: Date.now(),
-          content
-        })
-      )
-    } catch (ex) {
-      if (!retry) {
-        // probably failed because there's not enough space
-        // delete up all the other entries and try again
-        this.cleanupAutoSave(true)
-        this.doAutoSave(e, true)
-      } else {
-        console.error('Autosave failed:', ex) // eslint-disable-line no-console
+      const content = editor.getContent({no_events: true})
+      try {
+        this.storage.setItem(
+          this.autoSaveKey,
+          JSON.stringify({
+            autosaveTimestamp: Date.now(),
+            content
+          })
+        )
+      } catch (ex) {
+        if (!retry) {
+          // probably failed because there's not enough space
+          // delete up all the other entries and try again
+          this.cleanupAutoSave(true)
+          this.doAutoSave(e, true)
+        } else {
+          console.error('Autosave failed:', ex) // eslint-disable-line no-console
+        }
       }
     }
   }
@@ -1040,6 +1035,7 @@ class RCEWrapper extends React.Component {
   handleTextareaChange = () => {
     if (this.isHidden()) {
       this.setCode(this.textareaValue())
+      this.doAutoSave()
     }
   }
